All attrs should be optional.
Different agencies don't provide different values, so it is best (for now) that
we just assume that everything is potentially nullable.

The API wont provide the non-existing values, but it isn't RESTful to have the
representation of objects changing all the time, so we solve the problem on our
end.